Added Delete project/task/event function, fixed some bugs
Implemented the delete function for project/task/event, and fixed some bugs related to the check of parentId existence added last time in the add functions (some testcases used dummy parentIds and are causing errors). Now all the testcases are passed. Another fix is caused by the Google Sheet: it does not distinguish between "" and null (fails some testcases on deleting a parent and remove the parentId on its children), so I decided that each time we iterate through the sheet, every empty parentId are interpreted as null. Now all the testcases are passed.