Alpha Test Version for the MVP
Build basic front end logic and code, fixing the problems in the communication between google apps script and html (the date object is very tricky to handle, it needs to be converted to ISOStrings before passing in the traffic, but need to be filled in the Google Sheet as a Date object. The ISOString transition will also leads to a time zone discrepancy. All detected bug fixed. I am now doing the alpha test.